fix(#86): Index 위치 수정 및 불필요한 그래프 삭제 - 과거 NavHost를 위해 사용하던 불필요한 그래프 삭제
